requirements.txt
langchain==0.3.25
langchain-groq==0.3.2
langchain-core==0.3.58
python-dotenv==1.1.0
